pyproject.toml
[project]
name = "fakenews"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.5",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.2",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.5",
    "newsapi-python>=0.2.7",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "pdfplumber>=0.11.7",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
    "textblob>=0.19.0",
    "transformers>=4.56.2",
    "uvicorn>=0.36.0",
    "watchdog>=6.0.0",
    "websockets>=15.0.1",
]
